pyproject.toml
[project]
name = "highlightai2api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "filetype>=1.2.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pycryptodome>=3.23.0",
    "requests>=2.32.4",
    "sse-starlette>=3.0.2",
    "uvicorn>=0.35.0",
]
